# docs/openapi-v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "乒乓球赛事编排与赛务管理系统 Demo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tournaments/{tournament_id}": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "summary": "Get Tournament",
        "operationId": "get_tournament_api_tournaments__tournament_id__get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tournaments"
        ],
        "summary": "Delete Tournament",
        "description": "删除赛事（级联删除分组/选手/球台/比赛，见 db.py 外键 ON DELETE CASCADE）。",
        "operationId": "delete_tournament_api_tournaments__tournament_id__delete",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments/{tournament_id}/entries": {
      "get": {
        "tags": [
          "entries"
        ],
        "summary": "List Entries",
        "operationId": "list_entries_api_tournaments__tournament_id__entries_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntryOut"
                  },
                  "title": "Response List Entries Api Tournaments  Tournament Id  Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/pair-doubles": {
      "post": {
        "tags": [
          "entries"
        ],
        "summary": "Pair Doubles",
        "operationId": "pair_doubles_api_tournaments__tournament_id__pair_doubles_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/confirm-roster": {
      "post": {
        "tags": [
          "entries"
        ],
        "summary": "Confirm Roster",
        "operationId": "confirm_roster_api_tournaments__tournament_id__confirm_roster_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmRosterResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/groups": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Get Groups",
        "operationId": "get_groups_api_tournaments__tournament_id__groups_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/auto-group": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Auto Group",
        "operationId": "auto_group_api_tournaments__tournament_id__auto_group_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/ungroup": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Ungroup",
        "operationId": "ungroup_api_tournaments__tournament_id__ungroup_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/groups/{group_id}/qualification": {
      "patch": {
        "tags": [
          "groups"
        ],
        "summary": "Set Group Qualification",
        "operationId": "set_group_qualification_api_tournaments__tournament_id__groups__group_id__qualification_patch",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          },
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupQualifyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments/{tournament_id}/generate-knockout": {
      "post": {
        "tags": [
          "knockout"
        ],
        "summary": "Generate Knockout",
        "operationId": "generate_knockout_api_tournaments__tournament_id__generate_knockout_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnockoutTree"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/knockout": {
      "get": {
        "tags": [
          "knockout"
        ],
        "summary": "Get Knockout",
        "operationId": "get_knockout_api_tournaments__tournament_id__knockout_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnockoutTree"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BronzeMode": {
        "type": "string",
        "enum": [
          "BRONZE_MATCH",
          "JOINT_BRONZE"
        ],
        "title": "BronzeMode"
      },
      "ConfirmRosterResult": {
        "properties": {
          "tournament": {
            "$ref": "#/components/schemas/TournamentOut"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "tournament",
          "entries"
        ],
        "title": "ConfirmRosterResult"
      },
...
      "EntryMemberOut": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "college": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College"
          },
          "rating_points": {
            "type": "integer",
            "title": "Rating Points"
          },
          "member_order": {
            "type": "integer",
            "title": "Member Order"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "name",
          "college",
          "rating_points",
          "member_order"
        ],
        "title": "EntryMemberOut"
      },
      "EntryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "entry_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "rating_points": {
            "type": "integer",
            "title": "Rating Points"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "seed_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed No"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/EntryMemberOut"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournament_id",
          "entry_type",
          "display_name",
          "rating_points",
          "group_id",
          "seed_no",
          "status",
          "members"
        ],
        "title": "EntryOut"
      },
      "EventType": {
        "type": "string",
        "enum": [
          "SINGLES",
          "DOUBLES"
        ],
        "title": "EventType",
        "description": "参赛项目。第一版完整支持单打与固定搭档双打。"
      },
...
      "GroupOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "qualify_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualify Count"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/GroupPlayerOut"
            },
            "type": "array",
            "title": "Players"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "type": "array",
            "title": "Entries",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sort_order",
          "players"
        ],
        "title": "GroupOut"
      },
      "GroupPlayerOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "college": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "college"
        ],
        "title": "GroupPlayerOut"
      },
      "GroupQualifyUpdate": {
        "properties": {
          "qualify_count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Qualify Count"
          }
        },
        "type": "object",
        "required": [
          "qualify_count"
        ],
        "title": "GroupQualifyUpdate"
      },
...
      "GroupingResult": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupOut"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "groups"
        ],
        "title": "GroupingResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnockoutRoundOut": {
        "properties": {
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/KnockoutMatchOut"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "round",
          "label",
          "matches"
        ],
        "title": "KnockoutRoundOut"
      },
      "KnockoutTree": {
        "properties": {
          "tournament": {
            "$ref": "#/components/schemas/TournamentOut"
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/KnockoutRoundOut"
            },
            "type": "array",
            "title": "Rounds"
          },
          "champion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "runner_up": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "placements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Placements",
            "default": []
          },
          "placement_matches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Placement Matches",
            "default": []
          },
          "champion_path_match_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Champion Path Match Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "tournament",
          "rounds",
          "champion",
          "runner_up"
        ],
        "title": "KnockoutTree"
      },
...
      "PairingRequest": {
        "properties": {
          "pairing_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pairing Seed"
          }
        },
        "type": "object",
        "title": "PairingRequest"
      },
      "PairingResult": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "type": "array",
            "title": "Entries"
          },
          "unpaired_players": {
            "items": {
              "$ref": "#/components/schemas/PlayerOut"
            },
            "type": "array",
            "title": "Unpaired Players"
          },
          "pairing_seed": {
            "type": "integer",
            "title": "Pairing Seed"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "unpaired_players",
          "pairing_seed"
        ],
        "title": "PairingResult"
      },
      "PlacementMode": {
        "type": "string",
        "enum": [
          "OFF",
          "COMPLETE",
          "TIERED"
        ],
        "title": "PlacementMode"
      },
      "PlayerBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "seed_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed No"
          },
          "member_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Member Names",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "PlayerBrief"
      },
...
      "PlayerOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "college": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "College"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "seed_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed No"
          },
          "rating_points": {
            "type": "integer",
            "title": "Rating Points",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournament_id",
          "name",
          "college",
          "group_id",
          "seed_no"
        ],
        "title": "PlayerOut"
      },
...
      "TournamentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "table_count": {
            "type": "integer",
            "title": "Table Count"
          },
          "group_count": {
            "type": "integer",
            "title": "Group Count"
          },
          "qualify_per_group": {
            "type": "integer",
            "title": "Qualify Per Group"
          },
          "stage": {
            "$ref": "#/components/schemas/TournamentStage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType",
            "default": "SINGLES"
          },
          "bronze_mode": {
            "$ref": "#/components/schemas/BronzeMode",
            "default": "JOINT_BRONZE"
          },
          "placement_mode": {
            "$ref": "#/components/schemas/PlacementMode",
            "default": "OFF"
          },
          "games_to_win": {
            "type": "integer",
            "title": "Games To Win",
            "default": 2
          },
          "points_to_win": {
            "type": "integer",
            "title": "Points To Win",
            "default": 11
          },
          "roster_confirmed": {
            "type": "boolean",
            "title": "Roster Confirmed",
            "default": false
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "date",
          "table_count",
          "group_count",
          "qualify_per_group",
          "stage",
          "created_at"
        ],
        "title": "TournamentOut"
      },
      "TournamentStage": {
        "type": "string",
        "enum": [
          "REGISTRATION",
          "GROUP_STAGE",
          "KNOCKOUT",
          "FINISHED"
        ],
        "title": "TournamentStage",
        "description": "赛事阶段。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}